refactor and test StopSignalListen
